[not upstreamable] mtd: spi-nor: Use just 1-1-1 page program
The proper fix is to introduce a per-flash fixup hook to correct the
address lines and op code, but I'm letting this for other time, because
the performance benefit is negligible.

The benefit of PP x-y-4 is very limited as compared to PP 1-1-1,
the bottle neck being the page program time itself.

SST26 memories support Page Program 1-4-4, but with the op code of
Page Program 1-1-4 (32h). So the SPI NOR core set the 1-1-4 mode
(32h opcode), but the SST26 flash expects the address on 4 lines.
The I/O lines have pull-ups, so the flash receives a 0xffffff address
which is outside of its memory boundary, thus the operation is ignored.

Linux mainline, as well as our previuos linux4sam-6.0 release, uses just
1-1-1 pp exactly for this reason: there are plenty of custom op codes for
Page Program x-y-4 and it is hard to find a reliable op code that could
apply to most memories.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>